+++ R/get_signalp.R
@@ -15,6 +15,8 @@
 #' @param trunc An integer value corresponding to the N-terminal truncation of input sequence, at default set to 70. By default, the predictor truncates each sequence to max. 70 residues before submitting it to the neural networks. If you want to predict extremely long signal peptides, you can try a higher value, or disable truncation completely by entering 0 (zero).
 #' @param splitter An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Defaults to 500. Change only in case of a server side error. Accepted values are in range of 1 to 2000.
 #' @param sleep A numeric indicating the pause in seconds between POST and GET server calls, at default set to 1s. Decreasing is not recommended.
+#' @param attempts Integer, number of attempts if server unresponsive, at default set to 2.
+#' @param progress Bolean, whether to show the progress bar, at default set to TRUE
 #' @param ... currently no additional arguments are accepted apart the ones documented bellow.
 #'
 #' @return  A data frame with columns:
@@ -71,6 +73,8 @@ get_signalp.character <- function(data,
                                   trunc = 70L,
                                   splitter = 500L,
                                   sleep = 3,
+                                  attempts = 2,
+                                  progress = TRUE,
                                   ...){
   if (missing(splitter)) {
     splitter <- 500L
@@ -121,6 +125,47 @@ get_signalp.character <- function(data,
     if (trunc == 0){
       trunc <- 1000000L
     }
+  }
+  if (length(attempts) > 1){
+    attempts <- 2L
+    warning("attempts should be of length 1, setting to default: attempts = 2",
+            call. = FALSE)
+  }
+  if (!is.numeric(attempts)){
+    attempts<- as.numeric(attempts)
+    warning("attempts is not numeric, converting using 'as.numeric'",
+            call. = FALSE)
+  }
+  if (is.na(attempts)){
+    attempts <- 2L
+    warning("attempts was set to NA, setting to default: attempts = 2",
+            call. = FALSE)
+  }
+  if (is.numeric(attempts)) {
+    attempts <- floor(attempts)
+  }
+  if (attempts < 1) {
+    attempts <- 2L
+    warning("attempts was set to less then 1, setting to default: attempts = 2",
+            call. = FALSE)
+  }
+  if (missing(progress)) {
+    progress <- TRUE
+  }
+  if (length(progress) > 1){
+    progress <- TRUE
+    warning("progress should be of length 1, setting to default: progress = TRUE",
+            call. = FALSE)
+  }
+  if (!is.logical(progress)){
+    progress <- as.logical(progress)
+    warning("progress is not logical, converting using 'as.logical'",
+            call. = FALSE)
+  }
+  if (is.na(progress)){
+    progress <- TRUE
+    warning("progress was set to NA, setting to default: progress = TRUE",
+            call. = FALSE)
   }
   if (missing(sleep)) {
     sleep <- 3
@@ -249,20 +294,23 @@ get_signalp.character <- function(data,
   url <- "http://www.cbs.dtu.dk/cgi-bin/webface2.fcgi"
   cfg_file <- "/usr/opt/www/pub/CBS/services/SignalP-4.1/SignalP.cf"
   file_list <- ragp::split_fasta(path_in = file_name,
-                                 path_out = "temp_signalp_",
+                                 path_out = "tmp_signalp_",
                                  num_seq = splitter,
                                  trunc = trunc)
   if(grepl("temp_", file_name)){
     unlink(file_name)
   }
-  for_pb <- length(file_list)
-  pb <- utils::txtProgressBar(min = 0,
-                              max = for_pb,
-                              style = 3)
+  if(progress){
+    for_pb <- length(file_list)
+    pb <- utils::txtProgressBar(min = 0,
+                                max = for_pb,
+                                style = 3)
+  }
   splt <- (seq_along(file_list) - 1) %/% 10
   file_list <- split(file_list,
                      splt)
-  res <- lapply(seq_along(file_list), function(k){
+  output <- vector("list", length(file_list)*10)
+  for(k in seq_along(file_list)){
     x <- file_list[[k]]
     jobid <- vector("character", 10)
     for (i in seq_along(x)) {
@@ -289,14 +337,16 @@ get_signalp.character <- function(data,
                                 ".//input[@name='jobid']")
       jobid[i] <- xml2::xml_attr(res,
                                  "value")
-      unlink(x[i])
-      utils::setTxtProgressBar(pb,
-                               floor(i/2) + (10 * (k - 1)))
-      Sys.sleep(sleep)
+      if(progress){
+        utils::setTxtProgressBar(pb,
+                                 floor(i/2) + (10 * (k - 1)))
+      }
+      #Sys.sleep(sleep)
     }
-
+    
     collected_res <- vector("list", length(x))
     for (i in seq_along(x)) {
+      time1 <- Sys.time()
       repeat {
         res2 <- httr::GET(url = url,
                           query = list(jobid = jobid[i],
@@ -324,20 +374,145 @@ get_signalp.character <- function(data,
                           as = "parsed"),
             ".//pre")
         )
-        res2_split <- unlist(strsplit(res2,
-                                      "\n"))
+        res2_split <- unlist(
+          strsplit(res2,
+                   "\n")
+        )
+        Sys.sleep(1)
         if (any(grepl("Cmax", res2_split))) {
           break
         }
+        
+        time2 <- Sys.time()
+        
+        max.time <- as.difftime(pmax(50, splitter),
+                                units = "secs")
+        
+        if ((time2 - time1) > max.time) {
+          res2_split <- NULL
+          if(progress) print(
+            paste("file",
+                  x[i],
+                  "took longer then expected")
+          )
+          break
+        }
       }
-      res2_split <- res2_split[(which(grepl("name", res2_split))[1] +
-                                  1):(which(grepl("/pre", res2_split ))[1] - 1)]
+      if (is.null(res2_split)) {
+        tms <- 0
+        while(tms < attempts && is.null(res2_split)){
+          if(progress) print(
+            paste("reattempting file",
+                  x[i])
+          )
+          file_up <-  httr::upload_file(x[i])
+          res <- httr::POST(url = url,
+                            encode = "multipart",
+                            body = list(configfile = cfg_file,
+                                        SEQSUB = file_up,
+                                        orgtype = org_type,
+                                        `Dcut-type` = Dcut_type,
+                                        `Dcut-noTM` = Dcut_noTM,
+                                        `Dcut-TM` = Dcut_TM,
+                                        graphmode = NULL,
+                                        format = "short",
+                                        minlen = minlen,
+                                        method = method,
+                                        trunc = as.character(trunc)))
+          res <- httr::content(res,
+                               as = "parsed")
+          res <- xml2::xml_find_all(res,
+                                    ".//input[@name='jobid']")
+          jobidi <- xml2::xml_attr(res,
+                                   "value")
+          
+          time1 <- Sys.time()
+          
+          repeat {
+            res2 <- httr::GET(url = url,
+                              query = list(jobid = jobidi,
+                                           wait = "20"))
+            bad <- xml2::xml_text(
+              xml2::xml_find_all(
+                httr::content(res2,
+                              as = "parsed"),
+                "//head")
+            )
+            if (grepl("Illegal", bad)) {
+              prt <- xml2::xml_text(
+                xml2::xml_find_all(
+                  httr::content(res2,
+                                as = "parsed"),
+                  "//li")
+              )
+              stop(paste0(prt,
+                          ". Problem in file: ",
+                          "temp_",
+                          i,
+                          ".fa"),
+                   call. = FALSE)
+            }
+            res2 <- as.character(
+              xml2::xml_find_all(
+                httr::content(res2,
+                              as = "parsed"),
+                ".//pre")
+            )
+            res2_split <- unlist(
+              strsplit(res2,
+                       "\n")
+            )
+            Sys.sleep(1)
+            if (any(grepl("Cmax", res2_split))) {
+              break
+            }
+            
+            time2 <- Sys.time()
+            
+            max.time <- as.difftime(pmax(100, splitter * 1.5),
+                                    units = "secs")
+            
+            if ((time2 - time1) > max.time) {
+              res2_split <- NULL
+              break
+            }
+          }
+          tms <- tms + 1
+        }
+      }
+      if (is.null(res2_split)){
+        output <- do.call(rbind,
+                          output)
+        output$is.signalp <- output$is.sp == "Y"
+        if(progress){
+          utils::setTxtProgressBar(pb,
+                                   for_pb)
+          close(pb)
+        }
+        warning(
+          print(paste("maximum attempts reached at",
+                      x[i],
+                      "returning finished queries")),
+          call. = FALSE)
+        return(output)
+      }
+      unlink(x[i])
+      
+      res2_split <- res2_split[(which(grepl("name",
+                                            res2_split))[1] +
+                                  1):(which(grepl("/pre",
+                                                  res2_split ))[1] - 1)]
+      
       if(any(grepl("hr", res2_split))){
-        res2_split <- res2_split[1:(which(grepl("<hr>", res2_split))[1] - 1)]
+        res2_split <- res2_split[1:(which(grepl("<hr>",
+                                                res2_split))[1] - 1)]
       }
-      res2_split <- strsplit(res2_split, " +")
-      res2_split <- do.call(rbind, res2_split)
-      res2_split <- as.data.frame(res2_split, stringsAsFactors = F)
+      res2_split <- strsplit(res2_split,
+                             " +")
+      res2_split <- do.call(rbind,
+                            res2_split)
+      res2_split <- as.data.frame(res2_split,
+                                  stringsAsFactors = F)
       colnames(res2_split) <- c("id",
                                 "Cmax",
                                 "Cmax.pos",
@@ -350,22 +525,24 @@ get_signalp.character <- function(data,
                                 "is.sp",
                                 "Dmaxcut",
                                 "Networks.used")
-      utils::setTxtProgressBar(pb,
-                               floor(i/2) + 5 + (10 * (k - 1)))
-      collected_res[[i]] <- res2_split
+      if(progress){
+        utils::setTxtProgressBar(pb,
+                                 floor(i/2) + 5 + (10 * (k - 1)))
+      }
+      output[[((k*10)-10)+i]] <- res2_split
     }
-    collected_res <- do.call(rbind,
-                             collected_res)
-    collected_res$is.signalp <- collected_res$is.sp == "Y"
-    return(collected_res)
   }
-  )
-  res <- do.call(rbind,
-                 res)
-  utils::setTxtProgressBar(pb,
-                           for_pb)
-  close(pb)
-  return(res)
+  if(progress){
+    utils::setTxtProgressBar(pb,
+                             for_pb)
+    close(pb)
+  }
+  
+  output <- do.call(rbind,
+                    output)
+  
+  output$is.signalp <- output$is.sp == "Y"
+  return(output)
 }
 
 #' @rdname get_signalp